Add padding and margins as needed for portfolio section layout.
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss Please enter the commit message for your changes. Lines starting
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss with 'On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss' will be ignored, and an empty message aborts the commit.
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss On branch page-styling
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss Your branch is ahead of 'origin/page-styling' by 2 commits.
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss Changes to be committed:
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss 	modified:   app/index.html
On branch page-styling
Your branch is ahead of 'origin/page-styling' by 2 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   app/index.html
	modified:   app/styles/main.scss 	modified:   app/styles/main.scss